+++ src/app/modules/CoreModule/models/config.model.ts
@@ -1,0 +1,12 @@
+export interface ConfigModel {
+    id: string;
+    login: string;
+    email: string;
+    password: string;
+}
+
+export interface ProjectInfoModel {
+    App: string;
+    Ver: string;
+    API_URL: string;
+}