+++ src/app/modules/CoreModule/services/gen-id.generator.ts
@@ -1,0 +1,6 @@
+export function* genID() {
+    let id = 1;
+    while(true) {
+        yield id++
+    }
+}