+++ src/app/modules/CoreModule/services/generator.factory.ts
@@ -1,0 +1,8 @@
+import { InjectionToken } from "@angular/core";
+import { GeneratorService } from "./generator.service";
+
+export const generatedString = new InjectionToken<string>('generatedString');
+
+export function GeneratorFactory(n: number): (generator: GeneratorService) => string {
+    return (generator: GeneratorService) => generator.generate(n);
+}