+++ src/app/modules/MaterialModule/material.module.ts
@@ -1,9 +1,10 @@
-import { NgModule } from "@angular/core";
+import {NgModule} from "@angular/core";
 import {MatSidenavModule} from '@angular/material/sidenav';
 import {MatButtonModule} from '@angular/material/button';
+import {MatMenuModule} from '@angular/material/menu';
 
 @NgModule({
-    exports: [MatSidenavModule, MatButtonModule],
+    exports: [MatSidenavModule, MatButtonModule, MatMenuModule],
     declarations: []
 })
 export class MaterialModule {}