+++ src/changelog.md
@@ -4,3 +4,5 @@
 - Components for modules and mockups.
 - HighlightDirective for SharedModule.
 - Root service - CartService.
+
+- Added drop down menu to sort products using pipe OrderByPipe